feat: converting django routes to fastapi endpoints
