+++ app/models/user.py
@@ -25,7 +25,7 @@ class User(UserBase):
     is_active: bool
 
     class Config:
-        orm_mode = True
+        from_attributes = True
 
 
 class UserInDB(User):
