+++ app/schema/schema.py
@@ -2,3 +2,7 @@ from pydantic import BaseModel
 
 class Sequence(BaseModel):
     sequence: str
+
+class Consensus(BaseModel):
+    sequence: str
+    window_size: int
