--- app/tools/dna_visualization/admin.py
@@ -1,3 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
