--- app/tools/dna_visualization/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class DnaVisualizationConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'dna_visualization'
