--- app/tools/dna_visualization/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('', views.dna_visualization, name='dna_visualization'),
-]
