--- app/tools/gc_content/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class GcContentConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'gc_content'
