--- app/tools/gc_content/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('', views.gc_content, name='gc_content'),
-]
