--- app/tools/motif_finder/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class MotifFinderConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'motif_finder'
