--- app/tools/multiple_alignment/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class MultipleAlignmentConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'multiple_alignment'
