--- app/tools/multiple_alignment/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('', views.multiple_alignment, name='multiple_alignment'),
-]
