--- app/tools/pairwise_alignment/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class PairwiseAlignmentConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'pairwise_alignment'
