--- app/tools/pairwise_alignment/urls.py
@@ -1,6 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('', views.pairwise_alignment, name='pairwise_alignment'),
-]
